Fix camera jerking when mouse entered window.
Conceptually the problem is that on the first frame where the mouse
enters the window, its position is near the edge, so if we try to move
the camera based on that mouse position then the view will jerk way too
far. But just skipping the camera movement on that first frame doesn't
seem to be enough, due to a weird problem that only some of us could
reproduce. For some reason we found that on the second frame in which
the mouse warp call reported success, the mouse position was near the
position at which it had entered the window, as if the first warp had
failed. Very odd, especially since we tried getting the mouse position
after the first warp and that showed that the mouse was indeed centered.

We're not sure why this happens, but since we've only observed the
effect to last for one more frame (i.e., 2 in total), we fixed it by
just ignoring the mouse position until there had been 2 frames with
successful warps.