Remove app.globalClock.
In control.py where we use it, just call ClockObject.getGlobalClock()
instead. That's what we did back in WaRTS. We had avoided this
previously in SMUSh because pylint warned about it, but pylint warns
about _everything_ from panda3d.core, so a while back we just turned off
pylint module-member checking for panda3d.core. This means that we can
now use that method without needing to suppress a pylint warning.