Add collision groups.
This effectively replicates what we used to have with collision
bitmasks, though the mechanics are slightly different with BulletPhysics
than stock Panda3D. The point is, we are now able to define which
objects collide with which other objects, or to be more precise, which
pairs of objects do _not_ collide.

The main purpose of this change is so that we can make a rule that the
player does not collide with bullets, because otherwise newly-created
bullets immediately collide with the player with odd results.

Fixes #3.